requirements.txt
# Librerie per il machine learning e l'intelligenza artificiale
torch
torchvision
transformers
sentencepiece
protobuf
opencv-python
datasets

# Elaborazione di immagini
Pillow
timm

# Elaborazione di testo e NLP
nltk
spacy

# Librerie per la manipolazione dei file
numpy
pandas

# OCR (Riconoscimento ottico dei caratteri)
pytesseract

# Librerie per la gestione degli audio
librosa
soundfile

# Test e validazione
pytest

# Utilità varie
scikit-learn
tqdm

# Librerie per Docker (facoltative, se necessarie)
docker
